solution of search algorithm

+++ 0x1E-search_algorithms/0-linear.c
@@ -19,7 +19,7 @@ int linear_search(int *array, size_t size, int value)
 		return (-1);
 	}
 
-	for (n = 0; n < size; n++)
+	for (n = 0; n < (int)size; n++)
 	{
 		printf("Value checked array[%u] = [%d]\n", n, array[n]);
 		if (value == array[n])
